requirements.txt
aiofiles==23.2.1
aiohttp==3.8.5
chex==0.1.83
jax==0.4.23
msgpack_numpy==0.4.8
msgpack_python==0.5.6
numpy==1.26.0
pandas==2.1.1
PyYAML==6.0.1
PyYAML==6.0.1
torch==2.1.0
tqdm==4.66.1
uvloop==0.17.0
wandb
tabulate